chore: 메인페이지 컴포넌트로 분리
